+++ pydantic/__init__.py
@@ -1,3 +1,3 @@
 """DEA metamodel pydantic models (generated)."""
 
-__metamodel_version__ = "0.11.0"
+__metamodel_version__ = "0.12.0"